refactor: proper placements of common
