+++ __tests__/notifications.ts
@@ -13,7 +13,6 @@ import {
   Notification,
   NotificationAttachment,
   NotificationAvatar,
-  NotificationType,
   User,
 } from '../src/entity';
 import { DataSource } from 'typeorm';
@@ -23,6 +22,7 @@ import { notificationFixture } from './fixture/notifications';
 import { subDays } from 'date-fns';
 import request from 'supertest';
 import { FastifyInstance } from 'fastify';
+import { NotificationType } from '../src/notifications';
 
 let app: FastifyInstance;
 let con: DataSource;
