+++ src/entity/notifications/NotificationPreferenceComment.ts
@@ -1,11 +1,11 @@
-import { ChildEntity, OneToOne, PrimaryColumn } from 'typeorm';
-import { NotificationPreferenceType } from './NotificationPreference';
-
-@ChildEntity(NotificationPreferenceType.Comment)
-export class NotificationPreferencePost {
-  @PrimaryColumn({ type: 'text' })
-  commentId: string;
-
-  @OneToOne(() => Comment, { lazy: true, onDelete: 'CASCADE' })
-  comment: Promise<Comment>;
-}
+// import { ChildEntity, ManyToMany, PrimaryColumn } from 'typeorm';
+// import { NotificationPreferenceType } from './NotificationPreference';
+//
+// @ChildEntity(NotificationPreferenceType.Comment)
+// export class NotificationPreferencePost {
+//   @PrimaryColumn({ type: 'text' })
+//   commentId: string;
+//
+//   @ManyToMany(() => Comment, { lazy: true, onDelete: 'CASCADE' })
+//   comment: Promise<Comment>;
+// }
