+++ src/entity/notifications/NotificationPreferenceSource.ts
@@ -1,15 +1,12 @@
-import { ChildEntity, ManyToOne, PrimaryColumn } from 'typeorm';
-import { NotificationPreferenceType } from './NotificationPreference';
-import { Source } from '../Source';
-
-@ChildEntity(NotificationPreferenceType.Source)
-export class NotificationPreferencePost {
-  @PrimaryColumn({ type: 'text' })
-  sourceId: string;
-
-  @ManyToOne(() => Source, (source) => source.posts, {
-    lazy: true,
-    onDelete: 'CASCADE',
-  })
-  source: Promise<Source>;
-}
+// import { ChildEntity, ManyToMany, PrimaryColumn } from 'typeorm';
+// import { NotificationPreferenceType } from './NotificationPreference';
+// import { Source } from '../Source';
+//
+// @ChildEntity(NotificationPreferenceType.Source)
+// export class NotificationPreferencePost {
+//   @PrimaryColumn({ type: 'text' })
+//   sourceId: string;
+//
+//   @ManyToMany(() => Source, { lazy: true, onDelete: 'CASCADE' })
+//   source: Promise<Source>;
+// }
