+++ src/entity/notifications/index.ts
@@ -1,4 +1,3 @@
-export * from './common';
 export * from './Notification';
 export * from './NotificationAvatar';
 export * from './NotificationAttachment';
