+++ src/notifications/builder.ts
@@ -5,7 +5,6 @@ import {
   Notification,
   NotificationAttachment,
   NotificationAvatar,
-  NotificationType,
   Post,
   PostType,
   Source,
@@ -20,6 +19,7 @@ import { markdownToTxt } from 'markdown-to-txt';
 import { NotificationBundle, Reference } from './types';
 import { NotificationIcon } from './icons';
 import { SourceMemberRoles } from '../roles';
+import { NotificationType } from './common';
 
 const MAX_COMMENT_LENGTH = 320;
